% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetSE.R
\name{summary.glmnetSE}
\alias{summary.glmnetSE}
\title{Summary Function for glmnetSE Objects}
\usage{
\method{summary}{glmnetSE}(object, output = "coef")
}
\arguments{
\item{object}{A object of the class glmnetSE.}

\item{output}{A character string indicating if the coefficients with inference statistics "coef", or the performance metric with confidence intervalls should be displayed "metric" - default is "coef".}
}
\value{
The output of a glmnetSE object or the performance metric if cross fold validation is used.
}
\description{
Print the coefficients with standard errors and confidence intervals of a glmnetSE object. The inference statistics are only available for the coefficients without shrinkage, because otherwise they are biased. If cross fold validation is applied the performance metric can be displayed.
}
\examples{
\dontrun{
# Estimate modell
glmnetSE.model <- glmnetSE(data=swiss,cf.no.shrnkg = c("Education"))

# OUTPUT coefficients
summary.glmnetSE(glmnetSE.model, output="coef")

# OUTPUT performance metric
summary.glmnetSE(glmnetSE.model, output="metric")
}
}
\keyword{glmnetSE}
\keyword{output}
\keyword{results}
\keyword{summary}
