% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetSE.R
\name{glmnetSE}
\alias{glmnetSE}
\title{glmnetSE: Add bootstrap SE to glmnet for selected coefficients (no shrinkage)}
\usage{
glmnetSE(
  data,
  cf.no.shrnkg,
  alpha = 1,
  method = "none",
  r = 100,
  nlambda = 100,
  seed = 0,
  family = "gaussian"
)
}
\arguments{
\item{data}{A data frame, tibble or matrix object with the outcome variable in the first column and the feature variables in the following columns.}

\item{cf.no.shrnkg}{A character string of the coefficients that will be interpreted, the inference statistic is of interest and therefore no shrinkage will be applied.}

\item{method}{A character string defining if no 10-fold cross validation "none", 10-fold cross validation and selecting the lambda at which the MSE is within one standard error of the smallest MSE  "10CVoneSE", or 10-fold cross validation and selecting the lambda at which the smallest MSE is achieved "10CVmin" - default is "none.}

\item{r}{Number of bootstrap repetitions -default is 100.}

\item{nlambda}{Number of tested lambda values - default is 100.}

\item{seed}{Seed set for the bootstrap sampling - default 0 which means no seed set.}

\item{family}{A character string representing the used model family either "gaussian" or "binomial" - default is "gaussian".}

\item{aplha}{Alpha value 1 for LASSO and 0 for Ridge.}
}
\value{
A data frame with all coefficients, confidence intervals and information about significance of the effect for the coefficients of interesst.
}
\description{
glmnet or cv.glmnet with bootstrap standard errors for selected coefficients without shrinkage
}
\examples{
\dontrun{
GAUSSIAN with no cross validation and coefficient of interest Education
glmnetSE(data=swiss,cf.no.shrnkg = c("Education"), alpha=1, method="none", r=100, seed = 123, family="gaussian")

BINOMIAL with 10-fold cross validation selecting the lambda at which the smallest MSE is achieved, 500 bootstrap repetitions and coefficient of interest Education and Catholic.
glmnetSE(data=swiss,cf.no.shrnkg = c("Education", "Catholic"), alpha=1, method="10CVmin", r=500, seed = 123, family="binomial")
}
}
\keyword{bootstrap}
\keyword{errors}
\keyword{glmnet}
\keyword{shrinkage}
\keyword{standard}
