% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetSE.R
\name{summary.glmnetSE}
\alias{summary.glmnetSE}
\title{Summary Function for glmnetSE Objects}
\usage{
\method{summary}{glmnetSE}(object, ...)
}
\arguments{
\item{object}{A object of the class glmnetSE.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
The output of a glmnetSE object and the performance metric if cross fold validation is used.
}
\description{
Print the coefficients with standard errors and confidence intervals of a glmnetSE object. The inference statistics are only available for the coefficients without shrinkage, because otherwise they are biased. If cross fold validation is applied the performance metric can be displayed.
}
\examples{
\dontrun{
# Estimate model
glmnetSE.model <- glmnetSE(data=swiss,cf.no.shrnkg = c("Education"))

# Display model output with summary
summary(glmnetSE.model)
}
}
\keyword{glmnetSE}
\keyword{output}
\keyword{results}
\keyword{summary}
